Add additional methods to D-Bus interface
Replaces usage of sigc::slot objects in the handler map with
std::function objects. Fixed as part of this change is an oversight
where the static map was initialised with a pointer to a specific
Minecraftd1 instance.

The individual method handlers are also replaced with a single
handleSimpleCommand handler that writes the given command string to the
pipe connected to standard input; argument binding is used in the
handler map initialisation to supply the command values for each method.